add tests/test_category.py
