+++ tests/test_category.py
@@ -1,0 +1,13 @@
+from src.category import Category
+
+
+def test_category_init(first_category: Category, second_category: Category) -> None:
+    assert first_category.name == "Smart"
+    assert first_category.description == "Смартфоны, как средство не только коммуникации"
+    assert len(first_category.products) == 2
+
+    assert first_category.category_count == 2
+    assert second_category.category_count == 2
+
+    assert first_category.product_count == 3
+    assert second_category.product_count == 3
