+++ tests/test_utils.py
@@ -10,18 +10,6 @@ def test_product_init(product: Product) -> None:
     assert product.description == "256GB, Серый цвет, 200MP камера"
 
 
-def test_category_init(first_category: Category, second_category: Category) -> None:
-    assert first_category.name == "Smart"
-    assert first_category.description == "Смартфоны, как средство не только коммуникации"
-    assert len(first_category.products) == 2
-
-    assert first_category.category_count == 2
-    assert second_category.category_count == 2
-
-    assert first_category.product_count == 3
-    assert second_category.product_count == 3
-
-
 def test_read_json() -> None:
 
     # Создаем тестовый JSON-файл
